Serialize the mock genserver's pid
Elixir 1.5 has a new warning against storing PIDs in modules - this is
almost always a bug, as PIDs are ephemeral so they don't make sense to
sit in .beam files; however, in Simpler, we write (and compile) a
module at run-time, while everything is running, so in this case
it is not a bug.

To hide this behaviour we now put the serialized PID in the module
attribute, which is probably cleaner anyway.